requirements.txt
openai
python-dotenv
streamlit
pydantic
typing
psycopg2-binary
bcrypt